requirements.txt
# Core Framework
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.8

# Database
sqlalchemy==2.0.15
alembic==1.11.1

# Web Automation
playwright==1.34.0
pytest-playwright

# ML/AI Stack
transformers>=4.28.1
torch==1.13.1
accelerate==0.19.0
bitsandbytes==0.39.0
sentencepiece==0.1.99
protobuf==3.20.3
langchain==0.0.184
langchain-huggingface
chromadb==0.3.23

# Utilities
python-multipart==0.0.6
tqdm==4.65.0
numpy==1.24.3
tokenizers>=0.14
backports.lzma


# The current setup is opening the playwrite browser and loading the huggingface models.

# Help me reorganize the code and make it systematic.

# Also check if the current implementation is using LLMs and LangChain to make decisions on how browser 